Move functionality of RandomChooser to WeightedSampler
